Created url2html module, which will be used to parse urls and save the content on disk as a html file

+++ newsgrabber.py
@@ -1,7 +1,7 @@
 import yaml
 import gmailtool
 import logging
-
+import os
 
 logger = logging.getLogger(__name__)
 
@@ -13,7 +13,7 @@ class ConfigNotFound(Exception):
     def __str__(self):
         return repr('Configuration for service ' + "'" + self.value + "'" + ' not found!')
 
-class Config():
+class Config:
 
     def __init__(self, config_file):
 
@@ -72,48 +72,55 @@ class GmailConfig(Config):
         return self.gmail
 
 
-def parse_messages(messages, tags):
+def parse_messages(messages, category, content_dir):
 
-    for id, message in messages.iteritems():
+    for msg_id, message in messages.iteritems():
         if message.is_multipart():
-            print 'message id %s is multipart' % id
+            print 'message id %s is multipart' % msg_id
             for payload in message.get_payload():
                 if payload.get_content_type() == 'text/html': # and id == '15096053accd1dfa':
                     data = payload.get_payload(decode=True)
-                    new_data = create_metadata(id, message, 'html', data, tags)
+                    new_data = create_metadata(msg_id, message, 'html', data, category)
                     print 'payload type is text/html'
-                    file_name = 'pelican/content/' + id + '.html'
-                    with open(file_name, 'w') as f:
-                        f.write(new_data)
+                    file_name = content_dir + msg_id + '.html'
+                    if not os.path.isfile(file_name):
+                        with open(file_name, 'w') as f:
+                            f.write(new_data)
+                    else:
+                        logger.warn('Skipping. File already exists: %s', file_name)
         else:
-            metadata = create_metadata(id, message, 'markdown')
+            metadata = create_metadata(msg_id, message, 'markdown')
             content = create_content(message, 'markdown')
-            print 'message id %s is text' % id
-            file_name = 'pelican/content/' + id + '.md'
+            print 'message id %s is text' % msg_id
+            file_name = content_dir + msg_id + '.md'
             print 'writing file %s' % file_name
-            with open(file_name, 'w') as f:
-                #f.write('Title')
-                f.write(metadata)
-                for c in content:
-                    f.write('\n<' + c + '>')
+            if not os.path.isfile(file_name):
+                with open(file_name, 'w') as f:
+                    #f.write('Title')
+                    f.write(metadata)
+                    f.write(message.get_payload(decode=True).strip())
+                    #for c in content:
+                    #    f.write('\n<' + c + '>')
+            else:
+                logger.warn('Skipping. File already exists: %s', file_name)
 
-def create_metadata(id, message, format, data=None, tags=None):
+def create_metadata(msg_id, message, fmt, data=None, category=None):
 
-    if format == 'markdown':
-        metadata = 'Title: %s\nDate: %s\n' % (message['Subject'].strip(), message['Date'].strip())
+    if fmt == 'markdown':
+        metadata = 'Title: %s\nDate: %s\nSlug: %s\n' % (message['Subject'].strip(), message['Date'].strip(), str(msg_id))
         return metadata
 
-    if format == 'html':
+    if fmt == 'html':
         index = data.find('<meta')
-        print 'index: %s' % index
-        slug = '<meta name="Slug" content="' + str(id) + '">'
-        tag = '<meta name="tags" content="' + tags + '">'
-        new_data = data[:index] + slug + tag + data[index:]
+        slug = '<meta name="Slug" content="' + str(msg_id) + '">'
+        categories = '<meta name="category" content="' + category + '">'
+        print categories
+        new_data = data[:index] + slug + categories + data[index:]
         return new_data
 
-def create_content(message, format):
+def create_content(message, fmt):
 
-    if format == 'markdown':
+    if fmt == 'markdown':
         con = message.get_payload().strip().split()
         content = [c for c in con if 'http' in c]
         return content
@@ -125,6 +132,10 @@ def main():
     try:
         config = GmailConfig('conf/newsgrabber.conf').get_gmail_config()
         gmail = gmailtool.Gmail(config['secrets_file'], config['storage_file'], config['scopes'])
+        con_dir = Config('conf/newsgrabber.conf')
+        con_dir.load_config()
+        content_dir = con_dir.get_config('pelican')['content_dir']
+        print content_dir
 
     except KeyError:
 
@@ -138,7 +149,7 @@ def main():
     for k in config['search']:
         label = config['search'][k]
         messages = gmail.fetch_messages(label['query'])
-        parse_messages(messages, label['tags'])
+        parse_messages(messages, label['category'], content_dir)
